Introduce shared album art and use on regular card
