+++ src/components/AlbumArt/index.ts
@@ -1,0 +1,1 @@
+export * from "./AlbumArt";
