+++ src/components/MediocreMassiveMediaPlayerCard/index.ts
@@ -1,3 +1,2 @@
 export * from "./MediocreMassiveMediaPlayerCard";
 export * from "./MediocreMassiveMediaPlayerCardEditor";
-export * from "./components";
